+++ lib/karafka/routing/builder.rb
@@ -2,9 +2,9 @@
 
 module Karafka
   module Routing
-    # Routes builder used as a DSL layer for drawing and describing routes
+    # Builder used as a DSL layer for building consumers and telling them which topics to consume
     # @example Build a simple (most common) route
-    #   draw do
+    #   consumers do
     #     topic :new_videos do
     #       controller NewVideosController
     #     end
@@ -12,29 +12,19 @@ module Karafka
     class Builder < Array
       include Singleton
 
-      # All those options can be set on the route level except topic. Topic setup is part of
-      # the builder DSL, so we don't set it as an option
-      (Route::ATTRIBUTES - %i[topic]).each do |option|
-        define_method option do |value|
-          @current_route.public_send :"#{option}=", value
-        end
+      # Builds and saves given consumer group
+      # @param group_id [String, Symbol] name for consumer group
+      # @yield Evaluates a given block in a consumer group context
+      def consumer_group(group_id, &block)
+        self << ConsumerGroup.new(group_id.to_s, &block)
       end
 
-      # Creates a new route for a given topic and evalues provided block in builder context
-      # @param topic [String, Symbol] Kafka topic name
-      # @param block [Proc] block that will be evaluated in current context
-      # @note Creating new topic means creating a new route
-      # @example Define controller for a topic
-      #   topic :xyz do
-      #     controller XyzController
-      #   end
-      def topic(topic, &block)
-        @current_route = Route.new
-        @current_route.topic = topic
-
-        instance_eval(&block)
-
-        store!
+      # @param topic_name [String, Symbol] name of a topic from which we want to consumer
+      # @yield Evaluates a given block in a topic context
+      def topic(topic_name, &block)
+        self << ConsumerGroup.new(topic_name) do |consumer_group|
+          topic(topic_name, &block).tap(&:build)
+        end
       end
 
       # Used to draw routes for Karafka
@@ -48,34 +38,7 @@ module Karafka
       #   end
       def draw(&block)
         instance_eval(&block)
-      end
-
-      private
-
-      # Stores current route locally after it was built and validated
-      def store!
-        @current_route.build
-        @current_route.validate!
-
-        self << @current_route
-
-        validate! :topic, Errors::DuplicatedTopicError
-        validate! :group, Errors::DuplicatedGroupError
-      end
-
-      # Checks that among all routes a given attribute value is unique
-      # @param attribute [Symbol] what routes attribute we want to check for uniqueness
-      # @param error [Class] error class that should be raised when something is wrong
-      def validate!(attribute, error)
-        map = each_with_object({}) do |route, amounts|
-          key = route.public_send(attribute)
-          amounts[key] = amounts[key].to_i + 1
-          amounts
-        end
-
-        wrong = map.find { |_, amount| amount > 1 }
-
-        raise error, wrong if wrong
+        freeze
       end
     end
   end