+++ lib/karafka/routing/router.rb
@@ -7,16 +7,16 @@ module Karafka
     # @note Since Kafka does not provide namespaces or modules for topics, they all have "flat"
     #  structure so all the routes are being stored in a single level array
     class Router
-      # @param topic [String] topic based on which we find a proper route
+      # @param topic_id [String] topic based on which we find a proper route
       # @return [Karafka::Router] router instance
-      def initialize(topic)
-        @topic = topic
+      def initialize(topic_id)
+        @topic_id = topic_id
       end
 
       # Builds a controller instance that should handle message from a given topic
       # @return [Karafka::BaseController] base controller descendant instance object
       def build
-        route.controller.new.tap { |ctrl| ctrl.route = route }
+        topic.controller.new.tap { |ctrl| ctrl.topic = topic }
       end
 
       private
@@ -24,9 +24,14 @@ module Karafka
       # @return [Karafka::Routing::Route] proper route details
       # @raise [Karafka::Topic::NonMatchingTopicError] raised if topic name does not match
       #   any route defined by user using routes.draw
-      def route
-        App.routes.find { |route| route.topic == @topic } ||
-          raise(Errors::NonMatchingRouteError, @topic)
+      def topic
+        App.consumers.each do |consumer_group|
+          consumer_group.topics.each do |topic|
+            return topic if topic.id == @topic_id
+          end
+        end
+
+        raise(Errors::NonMatchingRouteError, @topic_id)
       end
     end
   end